mako: touch: define a power source for charging info
Define a touch power source and add it as a
supplicant for usb, dc and wireless power suppliers
so that the external_power_changed callback is
called when there is a power supply change.

When there is a power source supplied, touch driver
must set a special bit in the touch controller
so that the touch firmware can use advanced algorithm
to minimize the noise affection.

Change-Id: Idb3fedef87a8f91e2b9d3682fb07f79ee4004251

Conflicts:
	drivers/input/touchscreen/Kconfig