PM: extend suspend_again mechanism to use partialresume
The old platform suspend_again callback overrides drivers' votes, such that if
it implemented and returns false, then we do not call the partialresume
handlers.  When it doesn't exists or returns true, then we also query the
registered drivers for consensus.

When a device resumes from suspend, the suspend/resume code invokes
partialresume to check to see if the set of wakeup interrupts all have matching
handlers. If this is not the case, the PM subsystem can continue to resume as
before.  If all of the wakeup sources have matching handlers, then those are
invoked in turn (and can block), and if all of them reach consensus that the
reason for the wakeup can be ignored, they say so to the PM subsystem, which
goes right back into suspend.

* Fixed merge conflict by changing order of includes

Change-Id: I33ba1d882c9f29ea7904bb1af8aca904761b0cda
Signed-off-by: Iliyan Malchev <malchev@google.com>
Signed-off-by: Sudokamikaze <sudokamikaze@protonmail.com>